requirements.txt
reflex>=0.6.0
motor>=3.3.0
pymongo>=4.6.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
